splitting interfaces
