+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
   "files.associations": {
     "iosfwd": "cpp",
-    "fstream": "cpp"
+    "fstream": "cpp",
+    "iostream": "cpp"
   }
 }
