+++ treno.h
@@ -1,0 +1,20 @@
+#ifndef TRENO_H
+#define TRENO_H
+ 
+#include "gift.h"
+
+#include <vector>
+#include <iostream>
+#include <fstream>
+#include <vector>
+#include <list>
+#include <string>
+#include <bits/stdc++.h>
+
+class Treno {
+  public:
+    int capacity_;
+    std::vector<Gift> gifts_;
+};
+
+#endif
